requirements.txt
openai
ollama
pygame
python-dotenv
aiohttp
asyncio